Task 1 - Base Structure Created
The html file is updated to include a header, a main body section, and a footer. Within the heady and footer we used placeholders with a title for the browser as well as copyright info about the business and its employees. In the main section of the html, we added the div container with the id for employeeContainer where we will store employee information.