Task 2 - Employee Cards Dynamic Addition
Create a function that creates an employee card and displays the employee information linked to the employee card class and id. The employee card is created using createElement and specifying the div element type which is linked to the employee card class and id that is listed in the html using setAttribute. Using innerHTML, we are able to display the employee name and position that is plugged in when the function is called. A button is created and linked to the html by its id. This addEventListener makes it so that when the button is clicked, the employee card is removed from the employee container and the browser.