Task 4 - Employee Card Removal with Event Bubbling
From the removal button step in task 2, we will add a stop.Propagation operation that will make it so that when the the remove employee button is clicked, it will not bubble to the container and print "employee section clicked." This will only remove the employee. Adding an event listener to the new employee section, when there is a click on an employee card, the message "employee section clicked" will be logged.