Task 5 - Inline Editing for Employee Cards
Edit preexisting and add code so that an edit button is presented and when pressed, allows the browser user to edit the employee name and position within the browser. They then have the option to save with the save button. This save button will exit them out of the edit mode and display the replaced employee info. The code also allows the user to enter the edit mode by double clicking the employee card and saving it as well.